:star: Optimize ui with nav controller
