+++ app/src/main/java/com/mycars/di/modules/main/CarDetailFragmentBuilder.kt
@@ -1,5 +1,6 @@
 package com.mycars.di.modules.main
 
+import com.mycars.base.di.FragmentScope
 import com.mycars.carsdetail.di.modules.CarDetailViewModelModule
 import com.mycars.carsdetail.fragments.CarDetailFragment
 import dagger.Module
@@ -11,5 +12,6 @@ abstract class CarDetailFragmentBuilder {
     @ContributesAndroidInjector(
         modules = [CarDetailViewModelModule::class]
     )
+    @FragmentScope
     abstract fun bindCarDetailFragment(): CarDetailFragment
 }
