+++ app/src/main/res/menu/menu_navigation.xml
@@ -3,8 +3,10 @@
 
     <item
         android:id="@+id/main"
-        android:title="" />
+        android:icon="@drawable/ic_menu_car"
+        android:title="@string/app.menu_home" />
     <item
-        android:id="@+id/favorites"
-        android:title="" />
+        android:id="@+id/map"
+        android:icon="@drawable/ic_menu_map"
+        android:title="@string/app.menu_map" />
 </menu>
