+++ app/src/main/res/navigation/main.xml
@@ -3,21 +3,21 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/main"
-    app:startDestination="@id/carListFragment"
+    app:startDestination="@id/destination_list"
     tools:ignore="UnusedNavigation">
 
     <fragment
-        android:id="@+id/carListFragment"
-        android:name="com.mycars.carslist.fragments.CarListFragment"
+        android:id="@+id/destination_list"
+        android:name="com.mycars.carshome.fragments.CarListFragment"
         android:label="CarListFragment"
         tools:layout="@layout/fragment_car_list">
         <action
             android:id="@+id/action_detail"
-            app:destination="@id/detailFragment" />
+            app:destination="@id/destination_detail" />
     </fragment>
 
     <fragment
-        android:id="@+id/detailFragment"
+        android:id="@+id/destination_detail"
         android:name="com.mycars.carsdetail.fragments.CarDetailFragment"
         tools:layout="@layout/fragment_car_detail">
         <argument
