+++ base/src/main/java/com/mycars/base/di/Scopes.kt
@@ -5,3 +5,7 @@ import javax.inject.Scope
 @Scope
 @Retention
 annotation class ActivityScope
+
+@Scope
+@Retention
+annotation class FragmentScope
