+++ carshome/src/main/java/com/mycars/carshome/models/CarWidgetItem.kt
@@ -1,0 +1,11 @@
+package com.mycars.carshome.models
+
+import androidx.annotation.DrawableRes
+
+data class CarWidgetItem(
+    val id: Int,
+    @DrawableRes val imageType: Int,
+    val type: String,
+    val coordinates: String,
+    val heading: String
+)
