+++ carshome/src/main/res/values/strings.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <string name="carhome.coordinates_format">\@%1$.2f, %2$.2f</string>
+    <string name="carhome.heading">Heading: %1$.2f</string>
+</resources>
