--- carslist/src/main/java/com/mycars/carslist/models/CarWidgetItem.kt
@@ -1,11 +1,0 @@
-package com.mycars.carslist.models
-
-import androidx.annotation.DrawableRes
-
-data class CarWidgetItem(
-    val id: Int,
-    @DrawableRes val imageType: Int,
-    val type: String,
-    val coordinates: String,
-    val heading: String
-)
