+++ settings.gradle
@@ -5,6 +5,6 @@ include ':launcher'
 include ':baseui'
 include ':carsnetwork'
 include ':carsdata'
-include ':carslist'
+include ':carshome'
 include ':carsui'
 include ':carsdetail'
